% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aitoffFn}
\alias{aitoffFn}
\title{Project coordinates from Galactic to Aitoff-Hemmer.}
\usage{
aitoffFn(l, b)
}
\arguments{
\item{l}{in radians}

\item{b}{in radians}
}
\value{
dataframe with transposed coords (x,y)
}
\description{
Useful for skymap plots.
}
\examples{
aitoffFn(0,0)
}
