% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brewingMarkdown.R
\name{brew_plot_chunks}
\alias{brew_plot_chunks}
\title{Create markdown chunks automatically.}
\usage{
brew_plot_chunks(listItems, fn, xm.groups, xm.skymapData)
}
\arguments{
\item{listItems}{list of items to generate chunks on}

\item{fn}{function to generate data for the listItems}

\item{xm.groups}{actual data}

\item{xm.skymapData}{additional data, here the skymap for the function \code{fn}.}
}
\value{
chunks to be included in the markdown via 'child' reference
}
\description{
Check the code in the notebook as it contains escapes confusing for generated HTML.
Data for chunks is created in parallel and chunks are generated based on the created data, then executed in-place with the `child` argument of the chunk.
}
\examples{
\dontrun{brewed.chunks = brew_plot_chunks(xm.groups_big[1]
,plotAitoffGalacticOverlayBig
,xm.groups_big
,xm.skymap)

# include all brewed chunks
```{r brewing_big_chunks, child = brewed.chunks}

```
}
}
