% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbconnection.R
\name{connect}
\alias{connect}
\title{Creates connection to DPCG database
To be reused in the subsequent class and markdown SQL chunks.}
\usage{
connect(
  hostname = "localhost",
  dbname = "surveys",
  port = "55431",
  user = "johnny_local",
  pwd
)
}
\arguments{
\item{hostname}{DB hostname or localhost if ssh tunnel is used}

\item{dbname}{Usually surveys for DPCG}

\item{port}{55431 for Postgres-XZ}

\item{user}{DB user, i.e. nienarto_local}

\item{pwd}{optional password. Not needed if .pgpass is up to date (and it should be!)}
}
\value{
DB connection object to be reused in SQL chunks and other calls
}
\description{
Creates connection to DPCG database
To be reused in the subsequent class and markdown SQL chunks.
}
\examples{
\dontrun{connect(user="nienarto_local", port=55431)}
}
