% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRunMonitoring.R
\name{create1DHistogram}
\alias{create1DHistogram}
\title{Creates 1D histogram plot based on Run data.}
\usage{
create1DHistogram(histData)
}
\arguments{
\item{histData}{data retrieved}
}
\value{
ggplot with 1D histogram
}
\description{
PLots 1D histogram based on the Run data fetched beforehand
}
\examples{
\dontrun{
# get data for a specific monitoring plot
```{sql connection = conn, output.var = "runStats", echo = FALSE}
select * from getrun1dhistogram(?inRunId, 'Stat G time: mean \(([a-zA-Z]+)')
```
# plot it.
entryRow<-runStats[1,]
seriesName<- entryRow[,"attribute"]
create1DHistogram(runStats)
}
}
