% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{create1DHistogramRaw}
\alias{create1DHistogramRaw}
\title{Cfreate area hisgoram with cumulatvie sum}
\usage{
create1DHistogramRaw(histData, histMetaData, plotType = "area")
}
\arguments{
\item{histData}{dataframe with x,val}

\item{histMetaData}{dataframe with all needed fields: aribute, xlabel , scale = "LINEAR"/"LOG", failed, count for sidplauy purposes}

\item{plotType}{type of the plot (scatter,area, column, ...)}
}
\value{
highcharter area histogram
}
\description{
Cfreate area hisgoram with cumulatvie sum
}
\examples{
\dontrun{
periodData = sosSet \%>\%
group_by(x = round(period,2)) \%>\%
summarise(val = n()) \%>\% ungroup()
create1DHistogramRaw(periodData,
 metaData =
  data.frame(attribute = "Period",
  xlabel = paste(cu7Name, "Period"),
  scale = "LINEAR",
  failed = NaN,
  count = length(sosSet[[1]])))

}
}
