% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRunMonitoring.R
\name{create2DHistogramFractions}
\alias{create2DHistogramFractions}
\title{Create a 2D histogram based on a Run data.}
\usage{
create2DHistogramFractions(run2Dmeta, run2Ddata, reverseY = FALSE)
}
\arguments{
\item{run2Dmeta}{metadata about the plot}

\item{run2Ddata}{dataframe with the plot data}

\item{reverseY}{if to reverse Y scale, i.e. for Magnitude}
}
\value{
highcharts 2D histogram plot
}
\description{
Also out of a problematic (javascript-wise) data where we DO HAVE small fractional problem on the x,y axis.
}
\examples{
\dontrun{
# Get both metadata and data via two separate queries
```{sql connection = conn, output.var = "run2Dmeta", echo = FALSE}
select * From getRun2DHistogramMeta(?inRunId,
  'Stat G err: mean VS val: st. dev. \[counts\] \(EE')
```
```{sql connection = conn, output.var = "run2Ddata", echo = FALSE}
select * From getRun2DHistogramDataIdx(?inRunId,
 'Stat G err: mean VS val: st. dev. \[counts\] \(EE')
```
```{r hist2d_g_stdev_vs_meanerr, message=FALSE}
# plot.
create2DHistogramFractions(run2Dmeta,run2Ddata)

```
}
}
