% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dbstrToNumArray}
\alias{dbstrToNumArray}
\title{Function to convert from a database-recovered string to an array of numbers}
\usage{
dbstrToNumArray(theString)
}
\arguments{
\item{theString}{The string containing the array of strings from DB array}
}
\value{
a numeric array with the components
}
\description{
Function to convert from a database-recovered string to an array of numbers
}
\examples{
dbstrToNumArray("{1,2,3}")
}
