% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dbstrToStrArray}
\alias{dbstrToStrArray}
\title{Function to convert from a database-recovered string to an array of strings}
\usage{
dbstrToStrArray(theString)
}
\arguments{
\item{theString}{The string containing the array of strings from DB array}
}
\value{
a string array with the components
}
\description{
Function to convert from a database-recovered string to an array of strings
}
\examples{
dbstrToStrArray("{a,b,c}")
}
