% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{dfHPX}
\alias{dfHPX}
\title{Healpix data
function to convert hpx into sq degs for legend.
steradian * (180/pi)^2 = area in deg^2
https://lambda.gsfc.nasa.gov/toolbox/tb_pixelcoords.cfm}
\format{
An object of class \code{data.frame} with 11 rows and 6 columns.
}
\usage{
dfHPX
}
\description{
Healpix data
function to convert hpx into sq degs for legend.
steradian * (180/pi)^2 = area in deg^2
https://lambda.gsfc.nasa.gov/toolbox/tb_pixelcoords.cfm
}
\keyword{datasets}
