% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{dpcgTheme}
\alias{dpcgTheme}
\title{dpcg ggplot2 default plot scheme}
\usage{
dpcgTheme()
}
\value{
simple scheme
}
\description{
dpcg ggplot2 default plot scheme
}
\examples{
{
require(ggplot2)
theme = dpcgTheme()
}
}
