% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exportClassificationResults}
\alias{exportClassificationResults}
\title{Export results to prefix||variType table and to cumulativeTable classification table and other means with a given name.
Contrary to SOS version, It completely replaces cumulativeTable with the selection given in the data frame.}
\usage{
exportClassificationResults(
  conn = conn,
  schema,
  dbTableNameExport,
  inData,
  scoreName = "score",
  cumulativeTable = "dr3_classification_export",
  prefix = "dr3_validated_"
)
}
\arguments{
\item{conn}{DB connection}

\item{schema}{DB schema to create table in}

\item{dbTableNameExport}{name of the table}

\item{inData}{dataframe with results: must include sourceid, varitype and score fields.}

\item{scoreName}{name of the score field in the data frame provided}

\item{cumulativeTable}{name of the cumulative table, where we append sourceid and variType.}

\item{prefix}{prefix of the name of the snapshot table}
}
\value{
TRUE if ok.
}
\description{
Export results to prefix||variType table and to cumulativeTable classification table and other means with a given name.
Contrary to SOS version, It completely replaces cumulativeTable with the selection given in the data frame.
}
\examples{
\dontrun{
exportClassificationResults(dbTableNameExport, classSet, scoreName = "scoreColumn")
}
}
