% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{foldTimeseries}
\alias{foldTimeseries}
\title{Folding timeseries function}
\usage{
foldTimeseries(period, times, values, errors, referenceTime, range = 1.5)
}
\arguments{
\item{period}{given period}

\item{times}{vector of obstimes}

\item{values}{vector of values}

\item{errors}{vector of errors}

\item{referenceTime}{reference time for folding}

\item{range}{range of the folded timeseries}
}
\value{
dataframe with folded indexes, phases, magnitudes and errors
}
\description{
Folding timeseries function
}
\examples{
\dontrun{
foldTimeseries(period, times, values, errors, referenceTime, range=1.5)
}
}
