% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotAitoffGalacticOverlay}
\alias{plotAitoffGalacticOverlay}
\title{Plot overlay distribution in a skyplot}
\usage{
plotAitoffGalacticOverlay(bkg, classGroup, xm.skymap)
}
\arguments{
\item{bkg}{background plot we overlay on}

\item{classGroup}{name of the primaryvartype to filter on (for classification.)}

\item{xm.skymap}{dataframe to plot, with alpha, delta, and primaryvartype fields}
}
\value{
ggplot skymap plot.
}
\description{
Plot overlay distribution in a skyplot
}
\examples{
\dontrun{
# brew chunks on the xm.groups for a massive markdown generation
brewed.chunks = brew_chunks(xm.groups_big,plotAitoffGalacticOverlay,xm.groups_small,xm.skymap)
}
}
