% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotAitoffGalacticOverlayBig}
\alias{plotAitoffGalacticOverlayBig}
\title{Plot overlay distribution in a skyplot}
\usage{
plotAitoffGalacticOverlayBig(bkg, classGroup, xm.skymap, hpxLevel = 8)
}
\arguments{
\item{bkg}{background plot we overlay on}

\item{classGroup}{name of the primaryvartype to filter on (for classification.)}

\item{xm.skymap}{dataframe to plot, with alpha, delta, and primaryvartype fields}

\item{hpxLevel}{hpx level (for legend)}
}
\value{
ggplot skymap plot.
}
\description{
Uses geom_scattermore for plotting.
}
\examples{
\dontrun{
# brew chunks on the xm.groups for a massive markdown generation
brewed.chunks = brew_chunks(xm.groups_big,plotAitoffGalacticOverlay,xm.groups_small,xm.skymap)
}
}
