% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotAitoffGalacticOverlayBigSingleType}
\alias{plotAitoffGalacticOverlayBigSingleType}
\title{Plot overlay distribution in a skyplot with a single class attribute}
\usage{
plotAitoffGalacticOverlayBigSingleType(bkg, className, xm.skymap, hpxLevel = 8)
}
\arguments{
\item{bkg}{background plot we overlay on}

\item{className}{name of the primaryvartype to filter on (for classification.)}

\item{xm.skymap}{dataframe to plot, with alpha, delta, and primaryvartype fields}
}
\value{
ggplot skymap plot overlayd over bacground plot
}
\description{
Uses geom_scattermore for plotting.
}
\examples{
\dontrun{
# brew chunks on the xm.groups for a massive markdown generation
brewed.chunks = brew_chunks(xm.groups_big,plotAitoffGalacticOverlay,xm.groups_small,xm.skymap)
}
}
