% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotCmdAndHR}
\alias{plotCmdAndHR}
\title{Plots CMD and HR diafgrams in a combo}
\usage{
plotCmdAndHR(inData, valueName, catalogName = NULL, pcolour = "red")
}
\arguments{
\item{inData}{dataframe with CMD/HR data}

\item{valueName}{name of the plots}

\item{catalogName}{unused for now. We could show literature ones here.}

\item{pcolour}{points colours of the set points.}
}
\value{

}
\description{
Plots CMD and HR diafgrams in a combo
}
\examples{
\dontrun{
plotCmdAndHR(sosSet, valueName = cu7Name, catalogName = NULL)
}

}
