% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotTs}
\alias{plotTs}
\title{Create faceted plot based on a timeseries data and period, if exists.}
\usage{
plotTs(ts.all, sosSet)
}
\arguments{
\item{ts.all}{timeseries data to plot}

\item{sosSet}{dataframe with period}
}
\value{
faceted plot with timeseries and folded timeseries if period exists
}
\description{
Create faceted plot based on a timeseries data and period, if exists.
}
\examples{
\dontrun{
plotTs(ts.all, sosSet)
}
}
