% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{togalactic}
\alias{togalactic}
\title{Converting from ecliptic to galactic coordinates in J1950}
\usage{
togalactic(ra, dec)
}
\arguments{
\item{ra}{alpha in radians}

\item{dec}{delata in radians}
}
\value{
data frame of (gl,gb)
}
\description{
Converting from ecliptic to galactic coordinates in J1950
}
\examples{
togalactic(pi/4,pi/4)
}
