% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exportResults}
\alias{exportResults}
\title{Export results to DB and other means with a given name}
\usage{
exportResults(
  conn = conn,
  schema,
  dbTableNameExport,
  inData,
  variType,
  cumulativeTable = "dr3_common_export",
  prefix = "dr3_validated_"
)
}
\arguments{
\item{conn}{DB connection}

\item{schema}{DB schema to create table in}

\item{dbTableNameExport}{name of the table}

\item{inData}{dataframe with results.}

\item{variType}{name of the type}

\item{cumulativeTable}{name of the cumulative table, where we append sourceid and variType.}

\item{prefix}{prefix of the name of the snapshot table}
}
\value{
TRUE if ok.
}
\description{
Export results to DB and other means with a given name
}
\examples{
\dontrun{
exportResults(dbTableNameExport, sosSet)
}
}
